refactor: merge inject_test.go into chatprompt_test.go
